_archive/misc/outputs/requirements.txt
Flask==3.0.0
anthropic==0.39.0
pypdf==5.1.0
python-docx==1.1.0
Werkzeug==3.0.1